# Improve performance of `(to|from)_utc_timestamp` functions