1.  reuse the instance of  `DateLUT.instance("UTC")`, no need to create new instance every time in the loop.

2. not use `column.insert` but to create column to a certain size,  and give value to its element in the loop.

And we have test the performance of the pr, the test table schema `(x DateTime)`, which has `10000000` rows,  and the test sql `select count(1) from test where to_utc_timestamp(x, 'Etc/GMT+1') > '1990-01-01 12:00:00' SETTINGS max_threads=1`.

Before this PR
```
1 row in set. Elapsed: 1.561 sec. Processed 10.00 million rows, 40.00 MB (6.40 million rows/s., 25.62 MB/s.)
1 row in set. Elapsed: 1.544 sec. Processed 10.00 million rows, 40.00 MB (6.48 million rows/s., 25.91 MB/s.)
1 row in set. Elapsed: 1.572 sec. Processed 10.00 million rows, 40.00 MB (6.36 million rows/s., 25.44 MB/s.)
```    

After this PR
```
1 row in set. Elapsed: 0.876 sec. Processed 10.00 million rows, 40.00 MB (11.42 million rows/s., 45.66 MB/s.)
1 row in set. Elapsed: 0.894 sec. Processed 10.00 million rows, 40.00 MB (11.19 million rows/s., 44.77 MB/s.)
1 row in set. Elapsed: 0.944 sec. Processed 10.00 million rows, 40.00 MB (10.59 million rows/s., 42.36 MB/s.)
```

### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Functions `to_utc_timestamp` and `from_utc_timestamp` are now about 2x faster.

### Documentation entry for user-facing changes

- [ ] Documentation is written (mandatory for new features)

<!---
Directly edit documentation source files in the "docs" folder with the same pull-request as code changes

or

Add a user-readable short description of the changes that should be added to docs.clickhouse.com below.

At a minimum, the following information should be added (but add more as needed).
- Motivation: Why is this function, table engine, etc. useful to ClickHouse users?

- Parameters: If the feature being added takes arguments, options or is influenced by settings, please list them below with a brief explanation.

- Example use: A query or command.
-->


> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/

---
### Modify your CI run:
**NOTE:** If your merge the PR with modified CI you **MUST KNOW** what you are doing
**NOTE:** Checked options will be applied if set before CI RunConfig/PrepareRunConfig step

#### Include tests (required builds will be added automatically):
- [ ] <!---ci_include_fast--> Fast test
- [ ] <!---ci_include_integration--> Integration Tests
- [ ] <!---ci_include_stateless--> Stateless tests
- [ ] <!---ci_include_stateful--> Stateful tests
- [ ] <!---ci_include_unit--> Unit tests
- [ ] <!---ci_include_performance--> Performance tests
- [ ] <!---ci_include_asan--> All with ASAN
- [ ] <!---ci_include_tsan--> All with TSAN
- [ ] <!---ci_include_analyzer--> All with Analyzer
- [ ] <!---ci_include_KEYWORD--> Add your option here

#### Exclude tests:
- [ ] <!---ci_exclude_fast--> Fast test
- [ ] <!---ci_exclude_integration--> Integration Tests
- [ ] <!---ci_exclude_stateless--> Stateless tests
- [ ] <!---ci_exclude_stateful--> Stateful tests
- [ ] <!---ci_exclude_performance--> Performance tests
- [ ] <!---ci_exclude_asan--> All with ASAN
- [ ] <!---ci_exclude_tsan--> All with TSAN
- [ ] <!---ci_exclude_msan--> All with MSAN
- [ ] <!---ci_exclude_ubsan--> All with UBSAN
- [ ] <!---ci_exclude_coverage--> All with Coverage
- [ ] <!---ci_exclude_aarch64--> All with Aarch64
- [ ] <!---ci_exclude_KEYWORD--> Add your option here

#### Extra options:
- [ ] <!---do_not_test--> do not test (only style check)
- [ ] <!---no_merge_commit--> disable merge-commit (no merge from master before tests)
- [ ] <!---no_ci_cache--> disable CI cache (job reuse)

#### Only specified batches in multi-batch jobs:
- [ ] <!---batch_0--> 1
- [ ] <!---batch_1--> 2
- [ ] <!---batch_2--> 3
- [ ] <!---batch_3--> 4