@Algunenano 2024-04-12
> use DateLUT.to_time_t to convert timestamp between different time zone would lead to bad performance, and the time difference of 2 time zone is definite. So we can use the input time value to add or subtract this determined value, instead of use DateLUT.to_time_t to convert every time.

The time difference between 2 timezones depend on the time. You can't save/cache the value at one point and assume it's valid for another. A simple example: Right now Europe/Madrid is at UTC+2. A couple of weeks ago it was at UTC+1.