@KevinyhZou 2024-04-12
OK. But I think that the time zone is not often changed, and in the function computing period, perhaps a few seconds，the time zone is not  likely to change.