@KevinyhZou 2024-04-15
Hi, I have removed the logical of cache/save the value of 2 different timezone, and improve the performance by reuse the instance of `DateLUT.instance("UTC")`, and have updated the PR. please take a review at this pr. @rschu1ze @Algunenano